Rename api calls to start with api
