+++ frontend/src/pages/CDA/utils/api_images.js
