+++ frontend/src/pages/CDA/utils/api_trackbar.js
