+++ frontend/src/pages/CMT/utils/api_misc.js
@@ -1,16 +1,5 @@
 import { API } from "../../../core/config";
 
-export const getItemType = async () => {
-  const resp = await fetch(`${API}/CMT/item_type`, {
-    method: "GET",
-    headers: {
-      Accept: "application/json",
-    },
-  });
-  const json = await resp.json();
-  return json;
-};
-
 export const getFolderName = async (item) => {
   const resp = await fetch(`${API}/CMT/ds_folders`, {
     method: "POST",
@@ -41,6 +30,18 @@ export const getAllModels = async () => {
     headers: {
       Accept: "application/json",
     },
+  });
+  const json = await resp.json();
+  return json;
+};
+
+export const getFileCount = async (item, folder) => {
+  const resp = await fetch(`${API}/CMT/ds_file_count`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ item: item, folder: folder }),
   });
   const json = await resp.json();
   return json;
